Fixed bug with avatar upload
The avatar would not properly update if there was more than one avatar in the canvas avatars directory. 
To combat this the if statement was moved out of the for loop. So that it will confirm the upload after the loop has inspected all of the responses from the avatar's object.